# BarToYou-WebApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BarToYou API",
        "description": "Documentación Swagger de la API de BarToYou",
        "contact": {
            "email": "soporte@bartoyou.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bartoyou/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Listar miembros",
                "operationId": "5b0455a3a013746c085a28f1a4a58482",
                "responses": {
                    "200": {
                        "description": "Lista paginada de miembros"
                    }
                }
            },
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Crear un nuevo miembro",
                "operationId": "47cd4a2b07df349dd05d5882ad7c8b08",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "usuario123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Miembro creado correctamente"
                    },
                    "400": {
                        "description": "El miembro ya existe"
                    },
                    "409": {
                        "description": "El nombre ya está registrado"
                    }
                }
            }
        },
        "/api/bartoyou/members/{id}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Obtener miembro por ID",
                "operationId": "5d8fbc42bf911b3f6be6bf760c930c3c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Miembro encontrado"
                    },
                    "404": {
                        "description": "Miembro no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Members"
                ],
                "summary": "Actualizar miembro",
                "operationId": "f786651931b88561c9123ea28d040fd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "usuario123_editado"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Miembro actualizado correctamente"
                    },
                    "404": {
                        "description": "Miembro no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Members"
                ],
                "summary": "Eliminar miembro",
                "operationId": "a74848982be2849fe565d6377bbe5d97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Miembro eliminado correctamente"
                    },
                    "404": {
                        "description": "Miembro no encontrado"
                    }
                }
            }
        },
        "/api/check-token": {
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Verificar validez de un token",
                "operationId": "ef8250c49c138cd7a2cd5cc19a6c301e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "abc123xyz"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token válido, miembro retornado"
                    },
                    "401": {
                        "description": "Token inválido o expirado"
                    }
                }
            }
        },
...
    },
...
}